app-src/requirements.txt
Flask==3.1.1
boto3==1.40.3
fpdf2==2.8.3
Jinja2==3.1.6
flask-openapi3
flask-openapi3-swagger
Markdown